% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
Prints 'Hello, world!'
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}
}
