% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powercontinuous.R
\name{power.wald.continuous}
\alias{power.wald.continuous}
\title{Title}
\usage{
power.wald.continuous(n, t, rho, sigma, beta, sign.level)
}
\arguments{
\item{n}{number of individuals in each group}

\item{t}{number of periods}

\item{rho}{within-individual correlation}

\item{sigma}{the variance of an individual}

\item{beta}{treatment effect in the alternative hypothesis}

\item{sign.level}{alpha}
}
\value{
p
}
\description{
Title
}
\examples{
power.wald.continuous(n=20,t=2,rho=0.95,sigma=0.05,beta=1,sign.level=0.05)
}
